Refactor: Move the CFG and graph stuff to its own module
